server: split code and improve ci performance
