+++ server/BUILD.bazel
@@ -53,6 +53,7 @@ go_library(
         "//privilege/conn",
         "//privilege/privileges",
         "//privilege/privileges/ldap",
+        "//server/internal/util",
         "//server/metrics",
         "//session",
         "//session/txninfo",
@@ -182,6 +183,7 @@ go_test(
         "//parser/mysql",
         "//parser/terror",
         "//planner/core",
+        "//server/internal/util",
         "//session",
         "//sessionctx",
         "//sessionctx/binloginfo",
