+++ server/conn_stmt_test.go
@@ -19,6 +19,7 @@ import (
 	"context"
 	"encoding/binary"
 	"fmt"
+	"github.com/pingcap/tidb/server/internal/util"
 	"testing"
 
 	"github.com/pingcap/tidb/expression"
@@ -219,7 +220,7 @@ func TestParseExecArgsAndEncode(t *testing.T) {
 		[]byte{0x0},
 		[]byte{mysql.TypeVarchar, 0},
 		[]byte{4, 178, 226, 202, 212},
-		newInputDecoder("gbk"))
+		util.newInputDecoder("gbk"))
 	require.NoError(t, err)
 	require.Equal(t, "测试", dt[0].(*expression.Constant).Value.GetValue())
 
@@ -229,7 +230,7 @@ func TestParseExecArgsAndEncode(t *testing.T) {
 		[]byte{0x0},
 		[]byte{mysql.TypeString, 0},
 		[]byte{},
-		newInputDecoder("gbk"))
+		util.newInputDecoder("gbk"))
 	require.NoError(t, err)
 	require.Equal(t, "测试", dt[0].(*expression.Constant).Value.GetString())
 }
