+++ server/conn_test.go
@@ -22,6 +22,7 @@ import (
 	"crypto/x509"
 	"encoding/binary"
 	"fmt"
+	util2 "github.com/pingcap/tidb/server/internal/util"
 	"io"
 	"path/filepath"
 	"strings"
@@ -77,7 +78,7 @@ func TestIssue33699(t *testing.T) {
 
 	var outBuffer bytes.Buffer
 	tidbdrv := NewTiDBDriver(store)
-	cfg := newTestConfig()
+	cfg := util2.NewTestConfig()
 	cfg.Port, cfg.Status.StatusPort = 0, 0
 	cfg.Status.ReportStatus = false
 	server, err := NewServer(cfg, tidbdrv)
@@ -297,7 +298,7 @@ func TestInitialHandshake(t *testing.T) {
 	store := testkit.CreateMockStore(t)
 
 	var outBuffer bytes.Buffer
-	cfg := newTestConfig()
+	cfg := util2.NewTestConfig()
 	cfg.Port = 0
 	cfg.Status.StatusPort = 0
 	drv := NewTiDBDriver(store)
@@ -626,7 +627,7 @@ func testDispatch(t *testing.T, inputs []dispatchInput, capability uint32) {
 
 	var outBuffer bytes.Buffer
 	tidbdrv := NewTiDBDriver(store)
-	cfg := newTestConfig()
+	cfg := util2.NewTestConfig()
 	cfg.Port, cfg.Status.StatusPort = 0, 0
 	cfg.Status.ReportStatus = false
 	server, err := NewServer(cfg, tidbdrv)
@@ -1149,7 +1150,7 @@ func TestShowErrors(t *testing.T) {
 func TestHandleAuthPlugin(t *testing.T) {
 	store := testkit.CreateMockStore(t)
 
-	cfg := newTestConfig()
+	cfg := util2.NewTestConfig()
 	cfg.Port = 0
 	cfg.Status.StatusPort = 0
 	drv := NewTiDBDriver(store)
@@ -1530,7 +1531,7 @@ func TestChangeUserAuth(t *testing.T) {
 	tk := testkit.NewTestKit(t, store)
 	tk.MustExec("create user user1")
 
-	cfg := newTestConfig()
+	cfg := util2.NewTestConfig()
 	cfg.Port = 0
 	cfg.Status.StatusPort = 0
 
@@ -1578,7 +1579,7 @@ func TestChangeUserAuth(t *testing.T) {
 func TestAuthPlugin2(t *testing.T) {
 	store := testkit.CreateMockStore(t)
 
-	cfg := newTestConfig()
+	cfg := util2.NewTestConfig()
 	cfg.Port = 0
 	cfg.Status.StatusPort = 0
 
@@ -1770,7 +1771,7 @@ func TestOkEof(t *testing.T) {
 
 	var outBuffer bytes.Buffer
 	tidbdrv := NewTiDBDriver(store)
-	cfg := newTestConfig()
+	cfg := util2.NewTestConfig()
 	cfg.Port, cfg.Status.StatusPort = 0, 0
 	cfg.Status.ReportStatus = false
 	server, err := NewServer(cfg, tidbdrv)
@@ -1833,7 +1834,7 @@ func TestExtensionChangeUser(t *testing.T) {
 
 	var outBuffer bytes.Buffer
 	tidbdrv := NewTiDBDriver(store)
-	cfg := newTestConfig()
+	cfg := util2.NewTestConfig()
 	cfg.Port, cfg.Status.StatusPort = 0, 0
 	cfg.Status.ReportStatus = false
 	server, err := NewServer(cfg, tidbdrv)
@@ -1944,7 +1945,7 @@ func TestAuthSha(t *testing.T) {
 
 	var outBuffer bytes.Buffer
 	tidbdrv := NewTiDBDriver(store)
-	cfg := newTestConfig()
+	cfg := util2.NewTestConfig()
 	cfg.Port, cfg.Status.StatusPort = 0, 0
 	cfg.Status.ReportStatus = false
 	server, err := NewServer(cfg, tidbdrv)
@@ -2018,7 +2019,7 @@ func TestProcessInfoForExecuteCommand(t *testing.T) {
 
 func TestLDAPAuthSwitch(t *testing.T) {
 	store := testkit.CreateMockStore(t)
-	cfg := newTestConfig()
+	cfg := util2.NewTestConfig()
 	cfg.Port = 0
 	cfg.Status.StatusPort = 0
 	drv := NewTiDBDriver(store)
