+++ server/internal/util/BUILD.bazel
@@ -1,0 +1,31 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
+
+go_library(
+    name = "util",
+    srcs = ["util.go"],
+    importpath = "github.com/pingcap/tidb/server/internal/util",
+    visibility = ["//server:__subpackages__"],
+    deps = [
+        "//config",
+        "//parser/charset",
+        "//parser/mysql",
+        "//server",
+        "//types",
+        "//util/logutil",
+        "@org_uber_go_zap//:zap",
+    ],
+)
+
+go_test(
+    name = "util_test",
+    timeout = "short",
+    srcs = ["util_test.go"],
+    embed = [":util"],
+    flaky = True,
+    shard_count = 8,
+    deps = [
+        "//sessionctx/stmtctx",
+        "//types",
+        "@com_github_stretchr_testify//require",
+    ],
+)
